requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0